dashboard-mvp/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
psycopg2-binary==2.9.10
python-multipart==0.0.20
jinja2==3.1.5
itsdangerous==2.2.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
redis==5.2.1
celery==5.4.0
requests==2.32.3
